// link2rail/projects/ekp/src/app/empty-wagon/api/generated/model/demand-location.ts
/**
 * Empty wagon API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CommercialLocation } from './commercial-location';
import { FreightWagonLocation } from './freight-wagon-location';


export interface DemandLocation { 
    commercialLocation: CommercialLocation;
    freightWagonLocation?: FreightWagonLocation;
}